feat: ContentFormatter 타입 구현
